@@ -11,95 +12,96 @@
     "prod": "ng build --configuration production --source-map=false",
     "tsc-build": "tsc --build",
     "clean": "tsc --build --clean",
-    "dev:ssr": "ng run comparisonSite:serve-ssr",
-    "serve:ssr": "node dist/comparisonSite/server/main.js",
-    "build:ssr": "ng build && ng run comparisonSite:server",
-    "prerender": "ng run comparisonSite:prerender"
+    "serve:ssr:comparisonsite": "node dist/comparisonsite/server/server.mjs"
   },
   "private": true,
   "dependencies": {
     "@angular-devkit/core": "^19.2.4",
     "@angular-slider/ngx-slider": "^19.0.0",
     "@angular/animations": "^19.2.3",
-    "@angular/cdk": "^19.2.5",
+    "@angular/cdk": "^19.2.6",
     "@angular/common": "^19.2.3",
     "@angular/compiler": "^19.2.3",
     "@angular/core": "^19.2.3",
     "@angular/forms": "^19.2.3",
     "@angular/localize": "^19.2.3",
-    "@angular/material": "^19.2.5",
+    "@angular/material": "^19.2.6",
     "@angular/platform-browser": "^19.2.3",
     "@angular/platform-browser-dynamic": "^19.2.3",
     "@angular/platform-server": "^19.2.3",
     "@angular/router": "^19.2.3",
     "@angular/ssr": "^19.2.4",
     "@aposin/ng-aquila": "^19.2.0",
-    "@ckeditor/ckeditor5-angular": "^7.0.0",
+    "@ckeditor/ckeditor5-angular": "^7.0.1",
     "@fortawesome/fontawesome-free": "^6.7.2",
     "@material/data-table": "^14.0.0",
     "@ng-bootstrap/ng-bootstrap": "^18.0.0",
     "@ngneat/tailwind": "^7.0.3",
     "@ngneat/until-destroy": "^10.0.0",
     "@schematics/angular": "^19.2.4",
-    "@tailwindcss/forms": "^0.5.2",
-    "@tailwindcss/typography": "^0.5.4",
-    "@types/http-errors": "^2.0.1",
-    "@typescript-eslint/eslint-plugin": "^8.24.1",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/http-errors": "^2.0.4",
+    "@typescript-eslint/eslint-plugin": "^8.27.0",
     "angular-archwizard": "^7.0.0",
     "angular-count-to": "^0.0.3",
     "angularx-flatpickr": "^7.3.0",
-    "bootstrap": "^5.2.1",
-    "core-js": "^3.24.1",
-    "express": "^4.18.2",
+    "bootstrap": "^5.3.3",
+    "core-js": "^3.41.0",
+    "express": "^4.21.2",
     "i": "^0.3.7",
     "lottie-web": "^5.12.2",
-    "ng2-pdf-viewer": "^9.1.2",
+    "ng2-pdf-viewer": "^9.1.5",
     "ngx-bootstrap": "^19.0.2",
     "ngx-csv": "^0.3.2",
     "ngx-device-detector": "^9.0.0",
     "ngx-dropzone-wrapper": "^14.0.0",
-    "ngx-markdown": "^19.1.0",
+    "ngx-markdown": "^19.1.1",
     "ngx-take-until-destroy": "^5.4.0",
     "npkill": "^0.12.2",
-    "npm": "^10.9.0",
-    "path": "",
-    "pdfjs-dist": "^4.0.189",
+    "npm": "^10.9.2",
+    "path": "^0.12.7",
+    "pdfjs-dist": "^4.10.38",
     "primeicons": "^7.0.0",
     "primeng": "^17.18.15",
-    "rfdc": "^1.3.0",
-    "rxjs": "~6.5.3",
-    "semantic-release": "^19.0.3",
-    "simplebar-angular": "^3.2.4",
-    "sweetalert2": "^11.7.27",
-    "swiper": "^11.2.5",
-    "ts-loader": "^9.3.1",
+    "rfdc": "^1.4.1",
+    "rxjs": "~7.8.2",
+    "semantic-release": "^19.0.5",
+    "simplebar-angular": "^3.3.0",
+    "sweetalert2": "^11.17.2",
+    "swiper": "^11.2.6",
+    "ts-loader": "^9.5.2",
     "uuid": "^9.0.1",
-    "webpack": "^5.74.0",
+    "webpack": "^5.98.0",
     "webpack-dev-server": "^4.15.2",
+    "xhr2": "^0.2.1",
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^19.2.4",
     "@angular/cli": "^19.2.4",
     "@angular/compiler-cli": "^19.2.3",
-    "@types/express": "^4.17.17",
+    "@rollup/plugin-typescript": "^12.1.2",
+    "@types/express": "^4.17.21",
     "@types/jasmine": "~4.0.3",
-    "@types/node": "^18.18.0",
-    "autoprefixer": "^10.4.8",
-    "browser-sync": "^3.0.0",
+    "@types/node": "^18.19.81",
+    "autoprefixer": "^10.4.21",
+    "browser-sync": "^3.0.3",
     "jasmine-core": "~4.3.0",
-    "karma": "~6.4.0",
+    "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.1.1",
-    "karma-coverage": "~2.2.0",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "postcss": "^8.4.16",
+    "postcss": "^8.5.3",
     "prettier": "2.7.1",
-    "stylelint": "^14.9.1",
+    "stylelint": "^14.16.1",
     "stylelint-config-standard": "^26.0.0",
-    "tailwindcss": "^3.1.8",
+    "tailwindcss": "^3.4.17",
     "tslib": "^2.8.1",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "^5.4.3"
+    "typescript": "^5.8.2",
+    "vite": "^6.2.2",
+    "vite-plugin-angular": "^0.0.0"
   }
 }
