@@ -18,10 +18,11 @@ import {
   computed,
   OnDestroy,
   CUSTOM_ELEMENTS_SCHEMA,
-  NO_ERRORS_SCHEMA
+  NO_ERRORS_SCHEMA,
+  Signal, signal
 } from "@angular/core";
-import { Observable, forkJoin } from "rxjs";
-import { map, pairwise, filter, throttleTime } from "rxjs/operators";
+import { Observable, forkJoin, of } from "rxjs";
+import { startWith,catchError  } from 'rxjs/operators';
 import { CdkVirtualScrollViewport } from "@angular/cdk/scrolling";
 import { RouterLink, RouterLinkActive } from '@angular/router';
 import { NxActionComponent, NxActionIconDirective } from '@aposin/ng-aquila/action';
@@ -175,6 +176,9 @@ export class ResultsComponent implements OnInit, OnChanges, OnDestroy  {
   
   view: string = "list";
  
+  clinicasSignal: Signal<any[]> = signal([]); // Inicializamos la signal
+  planesSignal: Signal<any[]> = signal([]);
+  empresasSignal: Signal<any[]> = signal([]);
 
 
   receiveMessage($event: string) {
@@ -303,6 +307,45 @@ return this.cadena
   {
     this.formDataInicial = this.crearFormularioInicial();
     this.buildForm();
+    this.loadData();
+  }
+  loadData(): void {
+    // Llamadas a los servicios para obtener datos y actualizar las señales
+    this.cotizacionService.getClinicas().subscribe({
+      next: (clinicas: any) => {
+        // Actualizamos el valor de la señal directamente
+        this.clinicasSignal = signal(clinicas || []);
+      },
+      error: (error: any) => {
+        console.error('Error fetching clinicas', error);
+        // En caso de error, asignamos un array vacío
+        this.clinicasSignal = signal([]);
+      },
+    });
+
+    this.cotizacionService.getPlanes().subscribe({
+      next: (planes: any) => {
+        // Actualizamos el valor de la señal directamente
+        this.planesSignal = signal(planes || []);
+      },
+      error: (error: any) => {
+        console.error('Error fetching planes', error);
+        // En caso de error, asignamos un array vacío
+        this.planesSignal = signal([]);
+      },
+    });
+
+    this.cotizacionService.getEmpresas().subscribe({
+      next: (empresas: any) => {
+        // Actualizamos el valor de la señal directamente
+        this.empresasSignal = signal(empresas || []);
+      },
+      error: (error: any) => {
+        console.error('Error fetching empresas', error);
+        // En caso de error, asignamos un array vacío
+        this.empresasSignal = signal([]);
+      },
+    });
   }
   ngAfterContentChecked(): void {
     this.cdr.detectChanges();
@@ -879,18 +922,7 @@ if(this.previousSelectedItems.length > this.selectedItems.length){
     this.isLoaded = false;
     this.showComparionSidebar()
     this.isSmallScreen
-    forkJoin([
-      this.cotizacionService.getClinicas(),
-      this.cotizacionService.getPlanes(),
-      this.cotizacionService.getEmpresas(),
-    ]).subscribe(([clinicas, planes, empresas]) => {
-      this.clinicas = clinicas;
-      this.dropdownClinica = this.clinicas;
-      this.selectedClinica = [];
-      this.secureProducts = planes;
-      this.planes = planes;
-      this.empresas = empresas;
-    });
+  
         // Al iniciar el componente, verificamos el estado actual de la pantalla
       
 
