@@ -7,7 +7,7 @@ const app = express();
 app.use(express.static(__dirname + '/dist/comparisonsite'));
 
 app.get('/*', function(req,res) {
-    res.sendFile(path.join(__dirname+'/dist/comparisonSite/index.html'));
+    res.sendFile(path.join(__dirname+'/dist/comparisonsite/browser/index.html'));
 });
 
 app.listen(process.env.PORT || 8080);
