@@ -1,3 +1,3 @@
-export { INTERNAL_PATHS, INTERNAL_ROUTES, ROUTES_PATHS } from './internal.routes';
+export { INTERNAL_ROUTES } from './internal.routes';
 
 export const EMPTY_STRING: string = `` as const;
