@@ -1,114 +1,39 @@
-export const ROUTES_PATHS = {
-	AUTH: {
-		DEFAULT: `auth`,
-		LOGIN: `sigin`,
-		REGISTER: `sigup`,
-	},
-	ASESORI: {
-		DEFAULT: `asesori`,
-		CREDIT: {
-			DEFAULT: `credit`,
-			RESULTS: `results`,
-		},
-		INSURANCE: {
-			DEFAULT: `insurance`,
-			RESULTS: `results`,
-		},
-		SALUD: {
-			DEFAULT: ``,
-			RESULTS: ``,
-			DETAILS: `details`,
-			COMPARE: `compare`,
-			EMPRESA: `prepaga`,
-
-		},
-	},
-	SERVER: {
-		ERROR_404: `404`,
-		ERROR_400: `400`,
-		ERROR_401: `401`,
-		ERROR_403: `403`,
-		ERROR_500: `500`,
-	},
-};
-
-export const INTERNAL_PATHS = {
-	/**
-	 * Auth
-	 */
-	AUTH_DEFAULT: `${ROUTES_PATHS.AUTH.DEFAULT}`,
-	AUTH_LOGIN: `${ROUTES_PATHS.AUTH.LOGIN}`,
-	AUTH_REGISTER: `${ROUTES_PATHS.AUTH.REGISTER}`,
-	/**
-	 * ASESORI
-	 */
-	
-	ASESORI_DEFAULT: `${ROUTES_PATHS.ASESORI.DEFAULT}`,
-	/**
-	 * ASESORI - CREDIT
-	 */
-	ASESORI_CREDIT_DEFAULT: `${ROUTES_PATHS.ASESORI.CREDIT.DEFAULT}`,
-	ASESORI_CREDIT_RESULTS: `${ROUTES_PATHS.ASESORI.CREDIT.RESULTS}`,
-	/**
-	 * ASESORI - INSURANCE
-	 */
-	ASESORI_INSURANCE_DEFAULT: `${ROUTES_PATHS.ASESORI.INSURANCE.DEFAULT}`,
-	ASESORI_INSURANCE_RESULTS: `${ROUTES_PATHS.ASESORI.INSURANCE.RESULTS}`,
-	/**
-	 *  * ASESORI - INSURANCE
-	 */
-	ASESORI_SALUD_DEFAULT: `${ROUTES_PATHS.ASESORI.SALUD.DEFAULT}`,
-	ASESORI_SALUD_RESULTS: `${ROUTES_PATHS.ASESORI.SALUD.RESULTS}`,
-	ASESORI_SALUD_DETAILS: `${ROUTES_PATHS.ASESORI.SALUD.DETAILS}`,
-	ASESORI_SALUD_COMPARE: `${ROUTES_PATHS.ASESORI.SALUD.COMPARE}`,
-	ASESORI_SALUD_EMPRESA: `${ROUTES_PATHS.ASESORI.SALUD.EMPRESA}`,
-
-
-	/**
-	 * Server
-	 */
-	SERVER_ERROR_404: `${ROUTES_PATHS.SERVER.ERROR_404}`,
-	SERVER_ERROR_401: `${ROUTES_PATHS.SERVER.ERROR_401}`,
-	SERVER_ERROR_400: `${ROUTES_PATHS.SERVER.ERROR_400}`,
-	SERVER_ERROR_403: `${ROUTES_PATHS.SERVER.ERROR_403}`,
-	SERVER_ERROR_500: `${ROUTES_PATHS.SERVER.ERROR_500}`,
-};
-
 export const INTERNAL_ROUTES = {
 	/**
 	 * Auth
 	 */
-	AUTH_LOGIN: `/${INTERNAL_PATHS.AUTH_DEFAULT}/${INTERNAL_PATHS.AUTH_LOGIN}`,
-	AUTH_REGISTER: `/${INTERNAL_PATHS.AUTH_DEFAULT}/${INTERNAL_PATHS.AUTH_REGISTER}`,
+	AUTH_DEFAULT: `auth`,
+	AUTH_LOGIN: `auth/login`,
+	AUTH_REGISTER: `auth/signup`,
 	/**
 	 * Server
 	 */
-	SERVER_ERROR_404: `/${INTERNAL_PATHS.SERVER_ERROR_404}`,
-	SERVER_ERROR_401: `/${INTERNAL_PATHS.SERVER_ERROR_401}`,
-	SERVER_ERROR_400: `/${INTERNAL_PATHS.SERVER_ERROR_400}`,
-	SERVER_ERROR_403: `/${INTERNAL_PATHS.SERVER_ERROR_403}`,
-	SERVER_ERROR_500: `/${INTERNAL_PATHS.SERVER_ERROR_500}`,
+	SERVER_ERROR_404: `404`,
+	SERVER_ERROR_401: `401`,
+	SERVER_ERROR_400: `400`,
+	SERVER_ERROR_403: `403`,
+	SERVER_ERROR_500: `500`,
 	/**
 	 * ASESORI
 	 */
-	ASESORI_DEFAULT: `/${INTERNAL_PATHS.ASESORI_DEFAULT}`,
+	ASESORI_DEFAULT: `asesori`,
 	/**
 	 * ASESORI - CREDIT
 	 */
-	ASESORI_CREDIT_DEFAULT: `/${INTERNAL_PATHS.ASESORI_CREDIT_DEFAULT}`,
-	ASESORI_CREDIT_RESULTS: `/${INTERNAL_PATHS.ASESORI_CREDIT_DEFAULT}/${INTERNAL_PATHS.ASESORI_CREDIT_RESULTS}`,
+	ASESORI_CREDIT_DEFAULT: `credit`,
+	ASESORI_CREDIT_RESULTS: `credit/results`,
 	/**
 	 * ASESORI - INSURANCE
 	 */
-	ASESORI_INSURANCE_DEFAULT: `/${INTERNAL_PATHS.ASESORI_INSURANCE_DEFAULT}`,
-	ASESORI_INSURANCE_RESULTS: `/${INTERNAL_PATHS.ASESORI_INSURANCE_DEFAULT}/${INTERNAL_PATHS.ASESORI_INSURANCE_RESULTS}`,
+	ASESORI_INSURANCE_DEFAULT: `insurance`,
+	ASESORI_INSURANCE_RESULTS: `insurance/results`,
 		/**
 	 * ASESORI - SALUD
 	 */
-	ASESORI_SALUD_DEFAULT: `/${INTERNAL_PATHS.ASESORI_SALUD_DEFAULT}`,
-	ASESORI_SALUD_RESULTS: `/${INTERNAL_PATHS.ASESORI_SALUD_DEFAULT}/${INTERNAL_PATHS.ASESORI_SALUD_RESULTS}`,
-	ASESORI_SALUD_DETAILS: `/${INTERNAL_PATHS.ASESORI_SALUD_DEFAULT}/${INTERNAL_PATHS.ASESORI_SALUD_RESULTS}/${INTERNAL_PATHS.ASESORI_SALUD_DETAILS}`,
-	ASESORI_SALUD_COMPARE: `/${INTERNAL_PATHS.ASESORI_SALUD_DEFAULT}/${INTERNAL_PATHS.ASESORI_SALUD_RESULTS}/${INTERNAL_PATHS.ASESORI_SALUD_COMPARE}`,
-	ASESORI_SALUD_EMPRESA: `/${INTERNAL_PATHS.ASESORI_SALUD_DEFAULT}/${INTERNAL_PATHS.ASESORI_SALUD_RESULTS}/${INTERNAL_PATHS.ASESORI_SALUD_EMPRESA}`,
+	ASESORI_SALUD_DEFAULT: `salud`,
+	ASESORI_SALUD_RESULTS: `salud/results`,
+	ASESORI_SALUD_DETAILS: `salud/results/details`,
+	ASESORI_SALUD_COMPARE: `salud/results/compare`,
+	ASESORI_SALUD_EMPRESA: `salud/results/prepaga`,
 
 };
