@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { INTERNAL_PATHS } from '../../data/constants/routes';
+import { INTERNAL_ROUTES } from '../../data/constants/routes';
 import { SigninComponent } from './pages/signin/signin.component';
 
 
