@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { INTERNAL_PATHS } from '../../data/constants/routes';
+import { INTERNAL_ROUTES } from '../../data/constants/routes';
 import { DefaultComponent } from './pages/default/default.component';
 import { ResultsComponent } from './pages/results/results.component';
 import { DetailsComponent } from './pages/details/details.component';
@@ -7,11 +7,9 @@ import { CompareComponent } from './pages/compare/compare.component';
 import { EmpresasComponent } from './pages/empresas/empresas.component';
 
 export const routes: Routes = [
-    { path: ``, component: ResultsComponent },
-    { path: INTERNAL_PATHS.ASESORI_SALUD_DEFAULT, component: DefaultComponent },
-    { path: INTERNAL_PATHS.ASESORI_SALUD_RESULTS, component: ResultsComponent },
-    { path: INTERNAL_PATHS.ASESORI_SALUD_DETAILS, component: DetailsComponent },
-    { path: INTERNAL_PATHS.ASESORI_SALUD_COMPARE, component: CompareComponent },
-    { path: INTERNAL_PATHS.ASESORI_SALUD_EMPRESA, component: EmpresasComponent },
-
+    { path: ``, component: DefaultComponent },
+    { path: INTERNAL_ROUTES.ASESORI_SALUD_RESULTS, component: ResultsComponent },
+    { path: INTERNAL_ROUTES.ASESORI_SALUD_DETAILS, component: DetailsComponent },
+    { path: INTERNAL_ROUTES.ASESORI_SALUD_COMPARE, component: CompareComponent },
+    { path: INTERNAL_ROUTES.ASESORI_SALUD_EMPRESA, component: EmpresasComponent },
 ];
